Coffee-Machine-sim upload
This is a code challenge from '100 Days of Python' course that I am currently taking in my free time. This code is written to create a simulation of a coffee machine program. This program asks what drink option you want, and then will subtract the ingredients from an internal log. The program also will keep track of funds and will calculate change for the customer, if necessary. The program is designed to only work with U.S. coinage.